requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
asyncio==4.0.0
Brotli==1.1.0
build==1.3.0
CacheControl==0.14.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cleo==2.1.0
click==8.2.1
colorama==0.4.6
crashtest==0.4.1
cssselect2==0.8.0
distlib==0.4.0
dnspython==2.7.0
docx2pdf==0.1.8
dulwich==0.22.8
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
fastjsonschema==2.21.2
ffmpeg-python==0.2.0
filelock==3.19.1
findpython==0.6.3
fire==0.7.1
fonttools==4.59.1
future==1.0.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
keyring==25.6.0
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
msgpack==1.1.1
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pathlib==1.0.1
pbs-installer==2025.8.18
pdf2docx==0.5.8
pillow==11.3.0
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry-core==2.1.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
pyee==13.0.0
Pygments==2.19.2
PyMuPDF==1.26.4
pyphen==0.17.2
pyproject_hooks==1.2.0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
sentry-sdk==2.35.2
shellingham==1.5.4
sniffio==1.3.1
starlette==0.47.3
termcolor==3.1.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tomlkit==0.13.3
tqdm==4.67.1
trove-classifiers==2025.8.6.13
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.32.0
watchfiles==1.1.0
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
zopfli==0.2.3.post1
zstandard==0.24.0
